sqlite_to_postgres/requirements.txt
python-dotenv==0.21.1
psycopg2-binary==2.9.5 